pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrumdo-mcp"
version = "0.1.0"
description = "MCP server for ScrumDo boards — connect any AI tool directly to your board"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "ScrumDo LLC", email = "support@scrumdo.com" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "scrumdo",
    "agile",
    "scrum",
    "kanban",
    "ai",
    "llm",
    "claude",
    "cursor",
    "windsurf",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://www.scrumdo.com"
Documentation = "https://github.com/ScrumDoLLC/MassSense/tree/main/mcp"
Repository = "https://github.com/ScrumDoLLC/MassSense"
"Bug Tracker" = "https://github.com/ScrumDoLLC/MassSense/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
scrumdo-mcp = "spryng_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["spryng_mcp"]
exclude = [
    "tests/",
    "test_live.py",
    "update_on914.py",
    ".env*",
]

[tool.hatch.build.targets.sdist]
exclude = [
    ".venv/",
    "tests/",
    "test_live.py",
    "update_on914.py",
    ".env*",
    "dist/",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
